refacto public creation form
