+++ opencti-platform/opencti-front/lang/back/de.json
@@ -126,7 +126,6 @@
   "Contextual only": "Nur kontextabhängig",
   "Count": "Zählung",
   "CPE": "CPE",
-  "Create public dashboard": "Öffentliches Dashboard erstellen",
   "Created": "Erstellt",
   "Created at": "Erstellt am",
   "Created by": "Erstellt von",
