+++ opencti-platform/opencti-front/lang/back/en.json
@@ -126,7 +126,6 @@
   "Contextual only": "Contextual only",
   "Count": "Count",
   "CPE": "CPE",
-  "Create public dashboard": "Create public dashboard",
   "Created": "Created",
   "Created at": "Created at",
   "Created by": "Created by",
