+++ opencti-platform/opencti-front/lang/back/es.json
@@ -126,7 +126,6 @@
   "Contextual only": "Sólo contextual",
   "Count": "Cuenta",
   "CPE": "CPE",
-  "Create public dashboard": "Crear un cuadro de mandos público",
   "Created": "Creado",
   "Created at": "Creado en",
   "Created by": "Creado por",
