+++ opencti-platform/opencti-front/lang/back/fr.json
@@ -126,7 +126,6 @@
   "Contextual only": "Contextuel uniquement",
   "Count": "Nombre",
   "CPE": "CPE",
-  "Create public dashboard": "Créer un tableau de bord public",
   "Created": "Créé",
   "Created at": "Créé à",
   "Created by": "Créé par",
