+++ opencti-platform/opencti-front/lang/back/ja.json
@@ -126,7 +126,6 @@
   "Contextual only": "コンテキストのみ",
   "Count": "偽陽性",
   "CPE": "CPE",
-  "Create public dashboard": "公開ダッシュボードの作成",
   "Created": "情報タイプ",
   "Created at": "で作成",
   "Created by": "作成者",
