+++ opencti-platform/opencti-front/lang/back/ko.json
@@ -126,7 +126,6 @@
   "Contextual only": "컨텍스트 전용",
   "Count": "개수",
   "CPE": "CPE",
-  "Create public dashboard": "공개 대시보드 만들기",
   "Created": "생성됨",
   "Created at": "생성된 시각",
   "Created by": "생성자",
