+++ opencti-platform/opencti-front/lang/back/zh.json
@@ -126,7 +126,6 @@
   "Contextual only": "仅根据上下文",
   "Count": "计数",
   "CPE": "用户端设备",
-  "Create public dashboard": "创建公共仪表板",
   "Created": "创建",
   "Created at": "创建于",
   "Created by": "创建者",
