+++ opencti-platform/opencti-front/src/private/components/workspaces/dashboards/publicDashboards/PublicDashboardCreation.tsx
@@ -1,13 +1,10 @@
 import Drawer, { DrawerControlledDialProps, DrawerVariant } from '@components/common/drawer/Drawer';
-import React, { Suspense, useEffect } from 'react';
-import PublicDashboardCreationForm, { dashboardsQuery } from '@components/workspaces/dashboards/publicDashboards/PublicDashboardCreationForm';
-import { useQueryLoader } from 'react-relay';
-import { PublicDashboardCreationFormDashboardsQuery } from '@components/workspaces/dashboards/publicDashboards/__generated__/PublicDashboardCreationFormDashboardsQuery.graphql';
+import React from 'react';
+import PublicDashboardCreationForm from '@components/workspaces/dashboards/publicDashboards/PublicDashboardCreationForm';
 import { RecordSourceSelectorProxy } from 'relay-runtime';
 import { PublicDashboardsListQuery$variables } from '@components/workspaces/dashboards/publicDashboards/__generated__/PublicDashboardsListQuery.graphql';
 import { useFormatter } from '../../../../../components/i18n';
 import useHelper from '../../../../../utils/hooks/useHelper';
-import Loader, { LoaderVariant } from '../../../../../components/Loader';
 import CreateEntityControlledDial from '../../../../../components/CreateEntityControlledDial';
 import { insertNode } from '../../../../../utils/store';
 
@@ -23,27 +20,6 @@ const PublicDashboardCreation = ({ paginationOptions }: PublicDashboardCreationP
   const { t_i18n } = useFormatter();
   const { isFeatureEnable } = useHelper();
 
-  const [dashboardsQueryRef, fetchDashboards] = useQueryLoader<PublicDashboardCreationFormDashboardsQuery>(dashboardsQuery);
-  const fetchDashboardsWithFilters = () => {
-    fetchDashboards(
-      {
-        filters: {
-          mode: 'and',
-          filterGroups: [],
-          filters: [{
-            key: ['type'],
-            values: ['dashboard'],
-          }],
-        },
-      },
-      { fetchPolicy: 'store-and-network' },
-    );
-  };
-
-  useEffect(() => {
-    fetchDashboardsWithFilters();
-  }, []);
-
   const updater = (store: RecordSourceSelectorProxy) => insertNode(
     store,
     'Pagination_publicDashboards',
@@ -58,20 +34,11 @@ const PublicDashboardCreation = ({ paginationOptions }: PublicDashboardCreationP
       controlledDial={isFeatureEnable('FAB_REPLACEMENT') ? PublicDashboardCreateDial : undefined}
     >
       {({ onClose }) => (
-        <>
-          {dashboardsQueryRef && (
-            <Suspense fallback={<Loader variant={LoaderVariant.inElement} />}>
-              <PublicDashboardCreationForm
-                updater={updater}
-                queryRef={dashboardsQueryRef}
-                onCancel={onClose}
-                onCompleted={() => {
-                  onClose();
-                }}
-              />
-            </Suspense>
-          )}
-        </>
+        <PublicDashboardCreationForm
+          updater={updater}
+          onCancel={onClose}
+          onCompleted={onClose}
+        />
       )}
     </Drawer>
   );
